simplify provider status
